Adding multiple friend invite support from homepage, with FB autocomplete per field!
User can click "+ Add friend" to get another input field on the page.
Each new input field comes with FB autocomplete.

The FB User id is now stored with every recipient, bomber, and friend
so we can start to associate back to FB accounts.

We'll probably want to centralize FB account ID links at some point, so
in the future we can keep track of which FB ids are associated with
which phone numbers.

This lets us do cool things like automatically show the faces (FB
profile photos) of all the people who were involved in putting together
a lovebomb, and making a nifty visualization out of it.